+++ web/packages/api/src/toEthereumSnowbridgeV2.ts
@@ -33,6 +33,8 @@ import { paraImplementation } from "./parachains"
 import { Context } from "./index"
 import { ETHER_TOKEN_ADDRESS, getAssetHubConversionPalletSwap } from "./assets_v2"
 import { getOperatingStatus } from "./status"
+import { CreateAgent } from "./registration/agent/createAgent"
+import { Wallet, TransactionReceipt } from "ethers"
 
 export { ValidationKind, signAndSend } from "./toEthereum_v2"
 
@@ -829,4 +831,27 @@ export const validateTransferFromParachain = async (
         },
         transfer,
     }
+}
+
+// Agent creation exports
+export type {
+    AgentCreation,
+    AgentCreationValidationResult,
+    AgentCreationInterface,
+} from "./registration/agent/agentInterface"
+
+export function createAgentCreationImplementation() {
+    return new CreateAgent()
+}
+
+export async function sendAgentCreation(
+    creation: any,
+    wallet: Wallet,
+): Promise<TransactionReceipt> {
+    const response = await wallet.sendTransaction(creation.tx)
+    const receipt = await response.wait(1)
+    if (!receipt) {
+        throw Error(`Transaction ${response.hash} not included.`)
+    }
+    return receipt
 }
